move reductions to SparseArray
