+++ sparse/_compressed/compressed.py
@@ -1,5 +1,4 @@
 import numpy as np
-import numba
 import operator
 from numpy.lib.mixins import NDArrayOperatorsMixin
 from functools import reduce
@@ -11,15 +10,12 @@ from .._coo.common import linear_loc
 from .._common import dot, matmul
 from .._utils import (
     normalize_axis,
-    equivalent,
     check_zero_fill_value,
     check_compressed_axes,
 )
 from .._coo.core import COO
 from .convert import uncompress_dimension, _transpose, _1d_reshape
 from .indexing import getitem
-
-_reduce_super_ufunc = {np.add: np.multiply, np.multiply: np.power}
 
 
 def _from_coo(x, compressed_axes=None):
@@ -708,322 +704,6 @@ class GCXS(SparseArray, NDArrayOperatorsMixin):
             return matmul(other, self)
         except NotImplementedError:
             return NotImplemented
-
-    def __array_ufunc__(self, ufunc, method, *inputs, **kwargs):
-        out = kwargs.pop("out", None)
-        if out is not None and not all(isinstance(x, GCXS) for x in out):
-            return NotImplemented
-
-        if getattr(ufunc, "signature", None) is not None:
-            return self.__array_function__(ufunc, (np.ndarray, GCXS), inputs, kwargs)
-
-        if out is not None:
-            kwargs["dtype"] = out[0].dtype
-
-        if method == "outer":
-            method = "__call__"
-
-            cum_ndim = 0
-            inputs_transformed = []
-            for inp in reversed(inputs):
-                inputs_transformed.append(inp[(Ellipsis,) + (None,) * cum_ndim])
-                cum_ndim += inp.ndim
-
-            inputs = tuple(reversed(inputs_transformed))
-
-        # if method == "__call__":
-        #    result = elemwise(ufunc, *inputs, **kwargs)
-        if method == "reduce":
-            result = GCXS._reduce(ufunc, *inputs, **kwargs)
-        else:
-            return NotImplemented
-
-        if out is not None:
-            (out,) = out
-            if out.shape != result.shape:
-                raise ValueError(
-                    "non-broadcastable output operand with shape %s "
-                    "doesn't match the broadcast shape %s" % (out.shape, result.shape)
-                )
-
-            out._make_shallow_copy_of(result)
-            return out
-
-        return result
-
-    @staticmethod
-    def _reduce(method, *args, **kwargs):
-        assert len(args) == 1
-
-        self = args[0]
-        if isinstance(self, ss.spmatrix):
-            self = GCXS.from_scipy_sparse(self)
-
-        return self.reduce(method, **kwargs)
-
-    def reduce(self, method, axis=(0,), keepdims=False, **kwargs):
-        """
-        Performs a reduction operation on this array.
-
-        Parameters
-        ----------
-        method : numpy.ufunc
-            The method to use for performing the reduction.
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to perform the reduction. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-        kwargs : dict
-            Any extra arguments to pass to the reduction operation.
-
-        Returns
-        -------
-        GCXS
-            The result of the reduction operation.
-
-        Raises
-        ------
-        ValueError
-            If reducing an all-zero axis would produce a nonzero result.
-
-        See Also
-        --------
-        numpy.ufunc.reduce : A similar Numpy method.
-        COO.reduce : Equivalent operation on COO arrays.
-        """
-        axis = normalize_axis(axis, self.ndim)
-        zero_reduce_result = method.reduce([self.fill_value, self.fill_value], **kwargs)
-        reduce_super_ufunc = None
-
-        if not equivalent(zero_reduce_result, self.fill_value):
-            reduce_super_ufunc = _reduce_super_ufunc.get(method, None)
-
-            if reduce_super_ufunc is None:
-                raise ValueError(
-                    "Performing this reduction operation would produce "
-                    "a dense result: %s" % str(method)
-                )
-
-        if axis is None:
-            if reduce_super_ufunc is None:
-                temp = method(self.data, self.fill_value)
-                val = method.reduce(temp, **kwargs)
-            else:
-                temp = method(
-                    self.data, reduce_super_ufunc(self.fill_value, self.size - self.nnz)
-                )
-                val = method.reduce(temp, **kwargs)
-            if keepdims:
-                out = np.array(val)
-                return out.reshape(np.ones(self.ndim, dtype=np.intp))
-            return val
-
-        if not isinstance(axis, tuple):
-            axis = (axis,)
-
-        r = np.arange(self.ndim, dtype=np.intp)
-        compressed_axes = [a for a in r if a not in set(axis)]
-        x = self.change_compressed_axes(compressed_axes)
-        idx = np.diff(x.indptr) != 0
-        indptr = x.indptr[:-1][idx]
-        indices = (np.arange(x._compressed_shape[0], dtype=np.intp))[idx]
-        data = method.reduceat(x.data, indptr, **kwargs)
-        counts = x.indptr[1:][idx] - x.indptr[:-1][idx]
-
-        result_fill_value = self.fill_value
-
-        if reduce_super_ufunc is None:
-            missing_counts = counts != x._compressed_shape[1]
-            data[missing_counts] = method(
-                data[missing_counts], self.fill_value, **kwargs
-            )
-        else:
-            data = method(
-                data,
-                reduce_super_ufunc(self.fill_value, x._compressed_shape[1] - counts),
-            ).astype(data.dtype)
-            result_fill_value = reduce_super_ufunc(
-                self.fill_value, x._compressed_shape[1]
-            )
-
-        # prune data
-        mask = ~equivalent(data, result_fill_value)
-        data = data[mask]
-        indices = indices[mask]
-        out = GCXS(
-            (data, indices, []),
-            shape=(x._compressed_shape[0],),
-            fill_value=result_fill_value,
-            compressed_axes=None,
-        )
-
-        out = out.reshape(tuple(self.shape[d] for d in compressed_axes))
-
-        if keepdims:
-            shape = list(self.shape)
-            for ax in axis:
-                shape[ax] = 1
-            out = out.reshape(shape)
-
-        if out.ndim == 0:
-            return out[()]
-
-        return out
-
-    def sum(self, axis=None, keepdims=False, dtype=None, out=None):
-        """
-        Performs a sum operation along the given axes. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to sum. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-        dtype: numpy.dtype
-            The data type of the output array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.sum` : Equivalent numpy function.
-        scipy.sparse.coo_matrix.sum : Equivalent Scipy function.
-        COO.sum : Equivalent operation on COO arrays.
-        """
-        return np.add.reduce(self, out=out, axis=axis, keepdims=keepdims, dtype=dtype)
-
-    def max(self, axis=None, keepdims=False, out=None):
-        """
-        Maximize along the given axes. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to maximize. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-        dtype: numpy.dtype
-            The data type of the output array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.max` : Equivalent numpy function.
-        scipy.sparse.coo_matrix.max : Equivalent Scipy function.
-        COO.max : Equivalent operation on COO arrays.
-        """
-        return np.maximum.reduce(self, out=out, axis=axis, keepdims=keepdims)
-
-    amax = max
-
-    def any(self, axis=None, keepdims=False, out=None):
-        """
-        See if any values along array are ``True``. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to minimize. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.all` : Equivalent numpy function.
-        COO.any : Equivalent operation on COO arrays.
-        """
-        return np.logical_or.reduce(self, out=out, axis=axis, keepdims=keepdims)
-
-    def all(self, axis=None, keepdims=False, out=None):
-        """
-        See if all values in an array are ``True``. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to minimize. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.all` : Equivalent numpy function.
-        COO.all : Equivalent operation on COO arrays.
-        """
-        return np.logical_and.reduce(self, out=out, axis=axis, keepdims=keepdims)
-
-    def min(self, axis=None, keepdims=False, out=None):
-        """
-        Minimize along the given axes. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to minimize. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-        dtype: numpy.dtype
-            The data type of the output array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.min` : Equivalent numpy function.
-        scipy.sparse.coo_matrix.min : Equivalent Scipy function.
-        COO.min : Equivalent operation on COO arrays.
-        """
-        return np.minimum.reduce(self, out=out, axis=axis, keepdims=keepdims)
-
-    amin = min
-
-    def prod(self, axis=None, keepdims=False, dtype=None, out=None):
-        """
-        Performs a product operation along the given axes. Uses all axes by default.
-
-        Parameters
-        ----------
-        axis : Union[int, Iterable[int]], optional
-            The axes along which to multiply. Uses all axes by default.
-        keepdims : bool, optional
-            Whether or not to keep the dimensions of the original array.
-        dtype: numpy.dtype
-            The data type of the output array.
-
-        Returns
-        -------
-        GCXS
-            The reduced output sparse array.
-
-        See Also
-        --------
-        :obj:`numpy.prod` : Equivalent numpy function.
-        COO.prod : Equivalent operation on COO arrays.
-        """
-        return np.multiply.reduce(
-            self, out=out, axis=axis, keepdims=keepdims, dtype=dtype
-        )
 
     def astype(self, dtype, casting="unsafe", copy=True):
         """
